Class Diagram
I'm gonna add my design documents to the GitHub to be able to look back on them. As for now I am working on them alongside my code. Updating it to any changes I make to check if they will stay or be further changed. I have to say, the class diagram is a mess right now. I need to update the word document as well.