Rooms Addition
I tried adding the rooms class and make it work with the map class to update the map inside the map grid. It was very hard and it took me a long time and a lot of trial and error, but I got one function working "automatically" right now. It needs to be improved, but I'll save it for now.