Room Randomizer and bug fixed
I added a room randomizer which only uses one room class. Fixed the bug with fillMap() function that would throw an error when generating to the right side. I forgot to consider that setGrid() would also add on top of the last iteration of fillMap() which was already at the border of the grid.