First Player Movement
This is the first case of a working display of the player movement following the map creation path.

I had to redo the whole setGrid function, creating new global variables (playerLocationNumber, arrayNumber, playerX and playerY), and update the old playerMovement() function as well as the roomCleared() function.

Tried to create an overload for the setGrid() but that wouldnt work, so I took it out. During the testing I created a playerClamping to avoid errors. It's not automatic tho, it tries to recreate what the fillMap() function does.