Rooms and Encounters/Enemy
Starting to work on the encounters that will be set up inside the rooms. For now I am planning on creating an encounter class that will define the encounters and then use the room class to determine which of the encounters it is. For this I also created the Enemy class that hasnt been defined. This will also lead to creating the fighting functions. Hopefully the encounter will start on each player movement, then delete it when the room is completed and the player is ready to go to the next room.