Encounters in process and class updates
I updated the encounters class adding functions and variables.

Updated the UI and added colours to it. Updated the wallCheck() to be to determine the type of room the player is in instead of returning a bool. For that I had to create a new function in the Room class that wouldnt have any chance in it to store the walls quickly.

Took out variables from the objects that I dont think will be needed from now on. May add them back in the future.

Change the amount of items the player deals with to 1 until the game is functional.

Will start doing a basic attack function to test that the encounters are indeed working.