Display/Encounter/Enemy/Stats update
Display is finish for now.

I will be working on the encounters now until they all have a base encounter done for each situation.
I worked a bit on the enemy class:
Now they "drop" a random amount of EXP when killed that is given to the player.
Now their stats change depending on the 'difficulty' variable in encounter which also affects their random stats.
The display properly show the enemy stats on the encounter section.

The display was slightly updated to have a description section on the encounter side.

Added an encounter to every room type; for now only enemy encounter and healing fountain do anything other than displaying the start of the encounter.

Stats were slightly updated, now there is a heal() function.

Fighting will be finished later on. For now there is only a working enemy damaging function. When working with the other encounters there will be a need to create a way to damage the player, that will be probably be used in the fighting.