Working version with basic encounters
This is a working version of the game that is still missing a winning condition. There is a winning bool included now to the function calling the game loop, still needs further developing as right now the game only finishes when the player dies.

It has all encounters to a basic level.

Slightly updated objects for the object generation function.

Some of the encounter class global variables were taken out as they could be done inside the functions or were unnecessary. That included the functions to work with them.

With the addition of a win bool in the game class, encounter now has a pointer to it to change it when the player dies and in the future, in the case of winning the game.

A roll function (rollD20()) has been added to show rolls during encounters to feel more involved during the encounters. It gives the amount required to get certain results and it also shows the roll result to the player. Will add special results for "critical failures"(1) and "critical succes"(20).

It needs refining and completely finishing the fighting mechanic as right now is very simple and very RNG dependant to win.

Display has some new functions added to display the encounter descriptions. The descriptionText() function was modified from a code taken from a page that will be given credit on the documents. The base code was used to divide a string every time a certain char was inside the string. I modified it to change line every 34 characters and when a certain characters is inside the string. It still needs fixing as sometimes adding the character to create a single space would add 2 like its counting both condition at the same time disregarding the reset of the variable counting the words so far.

Display had some other changes doing with moving some of the functions that were being called inside the encounters to be automatically called inside the function displaying the description that was being used inside the encounters giving it the number to specify which encounter it was. Needs some cleaning.

Took out the extra #includes in all the solutions.

Fixed the HP display as it was backwards, as well as the getDODGE() function that was not giving the right amount.

I will merge this version with main before doing further updates.

The game still needs updating, still in testing stages. I just haven't updated main with a working version in a while. I feel like I should've been updating it with each update, my mistake. Not anymore.

P.S. Dont try to delete a pointer twice, I went around for a bit confused why my game function was giving me errors when closing. I forgot that if I delete the win bool heap in encounter, I don't need to do it in the function as well... It took me a second to realise I was doing it twice. Now every time I see a "reading violation" or something like that, I'll think of an extra delete.