Display/Enemy/Map/Player/Stats updates
Added a new function in the Display class to handle the size of the text on the display section. Avoids going over the limits.

Nerfed enemies to make it easier to win the game.

Updated map to make automatic some of the changes in path.
Changed how fillMap() would handle the 'totalRooms' addition so that every time the function is called it would subtract the extra room generated. This way I take out the extra process of running an if to do a check, still works.
Now there is a on-work display of the current room of the player and the total amount of rooms.
Added that when the player has completed all rooms, the winning condition turns to true. So the player now wins once all the rooms are cleared.

Updated player level up to keep the 'currentHP' even when the player is over the max HP. Basically, further supported the mechanic of overhealing.

Fixed the displayLine() error in HP display, it was backwards, as well as changing the default 'DMG' to be equal to 'STR'.