Introduction, minor changes, commenting & refactor
Added a new introduction to the game loop.

Took out some functions that were not being used.

Updated some functions that were uncompleted.

Added a pause in between all actions.

Added the missing descriptions section for encounters.

IN PROCESS:
Did some refactoring with the code as well as checking that the layout was consistent and easy to read.

Added and updated comments.